python/requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi>=2023.7.22
charset-normalizer==3.0.1
idna==3.4
PyYAML==6.0
requests>=2.31.0
soupsieve==2.3.2.post1
urllib3>=1.26.18
